# Only decode some smbios data