We dont need to have all the available data from the smbios, we can just
send the needed parts.

The handles are all fixed via the SMBIOS reference specification[0]

[0] https://www.dmtf.org/sites/default/files/standards/documents/DSP0134_3.6.0.pdf

Signed-off-by: Itxaka <igarcia@suse.com>