@Itxaka 2022-08-11
> Ok as a workaround. This probably needs some testing about the time overhead of spawing dmicode six times instead of one.


Doesnt dmidecode just reads and parses /sys/firmware/dmi/tables/DMI ? The usual cache mechanisms (disk, memory) should hit here and the runtime be negligent no? can you run `time dmidecode` on a big server several times to see how much time it takes the first time VS the 6th?