@Itxaka 2022-08-11
on my local tests (but this is a normal tower) it actually takes more to run the full command that to run with the type flag.

```
zeus:~ # time dmidecode
.....
real	0m0.014s
user	0m0.002s
sys	0m0.000s
```

```
zeus:~ # time dmidecode -t 12
.....
real	0m0.001s
user	0m0.000s
sys	0m0.001s
```

once the cache is hot, further calls to the full dmidecode now bring the time down, but the first one is the slow one.