@Itxaka 2022-08-11
with a simple benchmark

```
func BenchmarkDecode(b *testing.B) {
	for i := 0; i < b.N; i++ {
		Decode()
	}
}
```


results (old is main, new is this branch)
```
name      old time/op  new time/op  delta
Decode-8  1.67ms ± 0%  1.68ms ± 0%  +0.26%  (p=0.015 n=9+8)
```

Again, this is on a normal tower, we would need a beefy server with a lot of cards and shit to test it properly